Agent_angel_server/requirements.txt
fastapi
uvicorn
websockets
playwright
google-generativeai
python-dotenv
watchfiles
Pillow
playwright-stealth
pydantic
requests
